requirements.txt
# 영상 처리 및 그래픽
moviepy==1.0.3
Pillow>=9.5,<11
numpy>=1.26 
opencv-python==4.8.1.78

# 웹 크롤링 & API 호출
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.100.0
selenium==4.15.0

# 데이터 처리 & 머신러닝
pandas>=2.2
scikit-learn==1.3.0
joblib==1.3.2
scipy>=1.12

# TTS 및 오디오 처리
gTTS==2.4.0
pydub==0.25.1
SpeechRecognition==3.10.0

# 자동화 & 스케줄링
schedule==1.2.0

# 웹 대시보드
streamlit==1.35.0
plotly==5.15.0
dash==2.14.1

# 텔레그램 봇
python-telegram-bot==20.4

# 유틸리티
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
click==8.1.7

# OpenAI API
openai==0.28.1

# 이미지 처리
imageio==2.31.5
imageio-ffmpeg==0.4.9

# 텍스트 처리
textblob==0.17.1
nltk==3.8.1


# 성능 모니터링
psutil==5.9.5
memory-profiler==0.61.0

# 웹드라이버 (Selenium)
webdriver-manager==4.0.1

# 암호화
cryptography==41.0.7

# 로깅
loguru==0.7.2

# 환경 설정
pyyaml==6.0.1
configparser==6.0.0

# HTTP 클라이언트
httpx<0.25,>=0.23

